Am modificat toate variabile ce retineau indecsi sa fie de tipul size_t. Am creat un constructor in Board() care sa initializeze tabla, am modificat logica functiilor din findWinner pentru tabla dinamica.

+++ Eter/Eter/player.cpp
@@ -128,7 +128,7 @@ void eter::Player::AddCardToHand(const Card& card)
 }
 
 
-bool Player::placeCard(int x, int y, const Card& card, Board& board)
+bool Player::placeCard(size_t x, size_t y, const Card& card, Board& board)
 {
 
 	auto it = std::find(m_cardsInHand.begin(), m_cardsInHand.end(), card);
